requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
boto3==1.7.22
botocore==1.10.22
certifi==2018.8.13
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-extensions==2.1.0
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
drf-yasg==1.10.0
flake8==3.5.0
future==0.16.0
gunicorn==19.8.1
idna==2.7
inflection==0.3.1
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mccabe==0.6.1
model-mommy==1.6.0
more-itertools==4.3.0
openapi-codec==1.3.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.6.0
py==1.5.4
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.7.1
pytest-django==3.4.1
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
ruamel.yaml==0.15.57
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
